Fix the signature of the Sweet constructor.

+++ core/src/mcol/treasuregame/assets/items/Sweet.java
@@ -5,7 +5,7 @@ import mcol.treasuregame.assets.Assets;
 public class Sweet extends Item {
 
     /** Constructor. */
-    public Sweet(int x, int y, int value) {
+    public Sweet(float x, float y, int value) {
         super(x, y, value, Assets.itemTexture[4][2]);
     }
 }
